fix: Update spec to produce Planning clients which are closer to master
Generated and master version are significantly different because the master clients don't specify most of the accepted fields. Instead the master client just pass kwargs to the server. The current schema also doesn't cover ALL possible server options because there are too many of them (they should be refactored) but it states good chunk of them.

In this commit fixed multiple small bugs in the schema, updated the mako template and rewrote the hardcoded part of the clients to be closer to other python clients (vision etc). In some I noticed that schema does not match the server side to I fixed it.